refactor: move projen postconfiguration to projenrc
